Revert "separate carcontroller and productcontroller"
